refactor: Improved organization, modularity, cleanup and init., and unused code
- Move RandomEventItem to surpriseexam package
- Extract GraveNumber enum from Grave class into separate class
- Extract RelationshipType enum from OSRSItemRelationshipSystem
- Remove unused PatternRelation enum (old system)
- Improved proper cleanup on #shutDown
- Added additional variable clearing when events have concluded such as #onNpcDespawned
- Remove unused dependencies and variables
- Use Guava Sets.newHashSet() for consistent collection initialization
- Add missing @Slf4j and @Singletonannotations to classes
	- Replace System.out.println with proper logging in OSRSItemRelationshipSystem
